Moved models to src/models
I moved admin review model for web to src/models which is now accessible using '@/models/...'

I also added a job model so we don't have to make interfaces for the job every time we handle job data.
